Move Surprise Bear script into druid spells
